% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XldpFunctions.R
\name{getSubmissionMasterTable}
\alias{getSubmissionMasterTable}
\title{Get Submission Master Table Data}
\usage{
getSubmissionMasterTable()
}
\value{
Dataframe containing subset of data from SubmissionMaster_WithDevelopedClaims
}
\description{
Submissions Master Table is updated every month by Brian Tran.
This table joins SNAP, CODS, XLDP, and other data together and gets unique policies.
Rather than reinvent the wheel, we'll use this to join to xldp warehouse data to cut out all the duplicate quotes
Note this table only pulls policies with effective dates within the past year of whenever this is run.
That is to facilitate determining inforce status.
}
