% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHelperFunctions.R
\name{left}
\alias{left}
\title{Left Function}
\usage{
left(string, char)
}
\arguments{
\item{char}{}
}
\value{

}
\description{
This function extracts the Left side of the input string by desired number of characters
}
\examples{
left("Hello World",5)
}
