% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHelperFunctions.R
\name{right}
\alias{right}
\title{Right Function}
\usage{
right(string, char)
}
\arguments{
\item{char}{}
}
\value{

}
\description{
This function extracts the right side of the input string by desired number of characters
}
\examples{
right("Hello World", 5)
}
